[ADD] storage: Delete release functionality
The file `internal/storage/queries.go` was modified to add a new method `DeleteRelease`. This method takes an `id` as a parameter and deletes the corresponding release from the database using the SQL query `DELETE FROM releases WHERE id = ?`. The method returns an error if the deletion fails, wrapping it with a descriptive message.