[ADD] internal: Support for selecting release commit and merge commit, and retrieving branches added
The files `internal/tui/model.go`, `internal/tui/update.go`, and `internal/tui/utils.go` were modified to support the selection between a release commit and a merge commit, as well as to obtain all branches in the project. In `internal/tui/model.go`, two new fields `releaseType` and `releaseBranch` were added to the `Model` struct. In `internal/tui/update.go`, new cases were added to handle "Release Commit" and "Merge Commit" actions, which set the `releaseType` and `releaseBranch` fields accordingly. The `createRelease` function was updated to use these new fields.

The file `internal/tui/utils.go` was updated to include a new function `GetGitBranches` which retrieves a list of all branches in the Git repository. This function uses the `git branch --list` command to get the list of branches and returns them as a slice of strings.