[ADD] internal: Release element actions update
The files `internal/tui/update.go` were modified to add actions when selecting a release element from the list. Two new actions were introduced: "Print in console" and "Copy to clipboard". When "Print in console" is selected, the release information is printed to the console. When "Copy to clipboard" is selected, the release information is copied to the clipboard using the `golang.design/x/clipboard` package. The `updateReleaseMainMenu` function was also updated to include these new actions in the menu.

The files `go.mod` and `go.sum` were updated to include new dependencies and version updates.