[REF] internal: Update Go dependencies for clipboard library
The library used for copying to the clipboard was modified. The file `go.mod` was updated to replace the dependency on `golang.design/x/clipboard` with `github.com/atotto/clipboard`. The file `go.sum` was also updated to reflect this change.

Additionally, the file `internal/tui/update.go` was modified to import the new library and update the code to use the new clipboard API. Specifically, the `clipboard.Write` function was replaced with `clipboard.WriteAll` and `tea.SetClipboard` was used to set the clipboard contents.

The file `internal/tui/update.go` was also modified to update the `Update` function to use the new clipboard API. The `finalMessage` variable was introduced to construct the message to be copied to the clipboard. The `tea.Sequence` function was used to set the clipboard contents and quit the application.