[ADD] internal: Add support for releasing new versions on Github
The implementation of releasing a version on Github was completed. The file `internal/storage/queries.go` was modified to add a new function `GetLatestRelease`, which retrieves the latest release for a given workspace. The file `internal/tui/update.go` was updated to handle the creation of a release in the repository, including uploading the release to Github. This involved adding new functions `execUploadRelease` and `UploadReleaseToGithub`, as well as modifying the `Update` function to handle the release creation process.

The files `internal/tui/statusbar/statusbar.go` and `internal/tui/utils.go` were also modified. The `internal/tui/statusbar/statusbar.go` file was updated to change the version number in the status bar from `v0.2.5` to `v0.2.6`. The `internal/tui/utils.go` file was updated to add new functions `ExtractJSONError` and `UploadReleaseToGithub`.